psyopsOS/requirements.txt
boto3
PyYAML
invoke
